requirements.txt
# Core
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# LLM
ollama>=0.1.0

# Databases
neo4j>=5.15.0
chromadb>=0.4.22

# Embeddings & NLP
sentence-transformers>=2.2.2
scispacy>=0.5.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# UI
streamlit>=1.30.0
pyvis>=0.3.2

# Testing
pytest>=7.4.0